requirements.txt
astroid==2.4.2
certifi==2020.11.8
chardet==3.0.4
idna==2.10
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysql-connector-python==8.0.22
protobuf==3.13.0
pylint==2.6.0
requests==2.24.0
six==1.15.0
toml==0.10.2
urllib3==1.25.11
wrapt==1.12.1
